feat: refactor edits and exchanges to have common flag toggle flow
