+++ packages/core/core-flows/src/order/steps/index.ts
@@ -17,6 +17,7 @@ export * from "./delete-order-change-actions"
 export * from "./delete-order-changes"
 export * from "./delete-order-shipping-methods"
 export * from "./exchange/cancel-exchange"
+export * from "./list-order-change-actions-by-type"
 export * from "./exchange/create-exchange"
 export * from "./exchange/create-exchange-items-from-actions"
 export * from "./exchange/delete-exchanges"
