+++ packages/core/core-flows/src/order/workflows/exchange/exchange-request-item-return.ts
@@ -33,7 +33,7 @@ import {
   throwIfOrderChangeIsNotActive,
 } from "../../utils/order-validation"
 import { createOrderChangeActionsWorkflow } from "../create-order-change-actions"
-import { computeAdjustmentsForPreviewWorkflow } from "../order-edit/compute-adjustments-for-preview"
+import { computeAdjustmentsForPreviewWorkflow } from "../compute-adjustments-for-preview"
 import { refreshExchangeShippingWorkflow } from "./refresh-shipping"
 
 /**
@@ -194,7 +194,13 @@ export const orderExchangeRequestItemReturnWorkflow = createWorkflow(
 
     const orderChange: OrderChangeDTO = useRemoteQueryStep({
       entry_point: "order_change",
-      fields: ["id", "status", "version", "carry_over_promotions"],
+      fields: [
+        "id",
+        "status",
+        "version",
+        "exchange_id",
+        "carry_over_promotions",
+      ],
       variables: {
         filters: {
           order_id: orderExchange.order_id,
@@ -317,17 +323,11 @@ export const orderExchangeRequestItemReturnWorkflow = createWorkflow(
       } as OrderDTO & { promotions: PromotionDTO[] }
     })
 
-    when(
-      { orderChange },
-      ({ orderChange }) => !!orderChange.carry_over_promotions
-    ).then(() => {
-      computeAdjustmentsForPreviewWorkflow.runAsStep({
-        input: {
-          order: orderWithPromotions,
-          orderChange,
-          exchange_id: orderExchange.id,
-        },
-      })
+    computeAdjustmentsForPreviewWorkflow.runAsStep({
+      input: {
+        order: orderWithPromotions,
+        orderChange,
+      },
     })
 
     const refreshArgs = transform(
