+++ packages/core/core-flows/src/order/workflows/index.ts
@@ -50,7 +50,7 @@ export * from "./mark-payment-collection-as-paid"
 export * from "./maybe-refresh-shipping-methods"
 export * from "./order-edit/begin-order-edit"
 export * from "./order-edit/cancel-begin-order-edit"
-export * from "./order-edit/compute-adjustments-for-preview"
+export * from "./compute-adjustments-for-preview"
 export * from "./order-edit/confirm-order-edit-request"
 export * from "./order-edit/create-order-edit-shipping-method"
 export * from "./order-edit/order-edit-add-new-item"
@@ -81,7 +81,7 @@ export * from "./return/update-receive-item-return-request"
 export * from "./return/update-request-item-return"
 export * from "./return/update-return"
 export * from "./return/update-return-shipping-method"
-export * from "./set-carry-over-promotion-flag"
+export * from "./on-carry-promotions-flag-set"
 export * from "./transfer/accept-order-transfer"
 export * from "./transfer/cancel-order-transfer"
 export * from "./transfer/decline-order-transfer"
