+++ packages/core/core-flows/src/order/workflows/order-edit/update-order-edit-item-quantity.ts
@@ -22,7 +22,7 @@ import {
   throwIfIsCancelled,
   throwIfOrderChangeIsNotActive,
 } from "../../utils/order-validation"
-import { computeAdjustmentsForPreviewWorkflow } from "./compute-adjustments-for-preview"
+import { computeAdjustmentsForPreviewWorkflow } from "../compute-adjustments-for-preview"
 import { fieldsToRefreshOrderEdit } from "./utils/fields"
 
 /**
